+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Changed
 
+- **Default `selectionLimit`** is now **`1`** (single-tap pick). Multi-select requires setting `selectionLimit` to a value greater than `1`. ([#43](https://github.com/fennelouski/SwiftUI-Web-Image-Picker/issues/43))
+
 ### Fixed
 
 ### Removed